+++ src/services/Providers.tsx
@@ -1,0 +1,20 @@
+'use client'
+
+import { QueryClientProvider } from '@tanstack/react-query'
+import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import { getQueryClient } from '@/services/get-query-client'
+
+import type * as React from 'react'
+import { Toaster } from 'sonner'
+
+export default function Providers({ children }: { children: React.ReactNode }) {
+	const queryClient = getQueryClient()
+
+	return (
+		<QueryClientProvider client={queryClient}>
+			<Toaster position="top-center" richColors />
+			{children}
+			<ReactQueryDevtools />
+		</QueryClientProvider>
+	)
+}